requirements.txt
openai==1.12.0
midiutil==1.2.1
pillow==10.2.0
python-dotenv==1.0.0